updated projects routes
