+++ src/views/layout.hbs
@@ -43,12 +43,12 @@
               </a>
             </li>
             <li class="nav-item">
-              <a onclick="toggleLoading()" class="nav-link {{#if projectsActive}}active{{/if}}" href="/projects">
+              <a onclick="toggleLoading()" class="nav-link {{#if projectsActive}}active{{/if}}" href="/stats">
                 <svg width="1em" height="1em" viewBox="0 0 16 16" class="bi bi-bar-chart-fill" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
                   <rect width="4" height="5" x="1" y="10" rx="1"/>
                   <rect width="4" height="9" x="6" y="6" rx="1"/>
                   <rect width="4" height="14" x="11" y="1" rx="1"/>
-                </svg> Projects
+                </svg> Projects' Stats
               </a>
             </li>
             <li class="nav-item">
