+++ src/views/stats.hbs
@@ -1,4 +1,4 @@
-<h2>Projects</h2>
+<h2>Projects' Stats</h2>
 
 <form id="filter">
   <input type="hidden" id="page" name="page" value="{{page}}"/>
