+++ src/views/track.hbs
@@ -52,8 +52,6 @@
         return a + `<option value="${b.id}">${b.name}</option>`
       }, '')
 
-      console.log(projectsOption)
-
       document.getElementById('project').innerHTML = projectsOption
     })
   }
